Re-package PirateLightClientKit for iOS
This change gives us complete control over the library, as opposed to relying on the upstream package manager.

One immediate benefit is that we can inject our own checkpoints, and control how they get packaged into the client app.

In the future this will allow us to perform package upgrades with more flexibility.