Stop using the iOS SQLite library
In order to work correctly, the Rust FFI code needs to link against SQLite >= 3.37.0. The way we satisfy this dependency is to include the SQLite.swift Pod, which in turn depends on the iOS-provided SQLite library. Unfortunately, iOS does not ship with a new enough SQLite until iOS 16, but we claim to support iOS 13.

To fix this problem, we need to bring our own version of SQLite, which is newer that the OS-provided one. Fortunately, the SQLite.swift library provides a way to do this - we just include the "/standalone" subspec.